cmake: move kokkos files into subdirectory
We end up with two cases:

A. Shared build
   realm_kokkos is a shared library and self-contained. Linking to it doesn't
   propagate link options.

B. Static build
   realm_kokkos is a static library and inherits the Kokkos INTERFACE link
   options. This forces us to use the same linker/wrapper as Kokkos in this
   case.
